requirements.txt
folium==0.19.4
streamlit==1.15.1
pandas==1.4.3
numpy==1.23.1
matplolib-inline==0.1.6
haversine==2.7.0
streamlit-folium==0.7.0
Pillow==9.2.0